+++ extra_platforms/__init__.py
@@ -202,7 +202,7 @@ from .operations import (  # noqa: E402
     reduce,
 )
 
-__version__ = "4.1.1"
+__version__ = "4.1.2"
 """Examples of valid version strings according :pep:`440#version-scheme`:
 
 .. code-block:: python